fix(import): scope media re-validation to imported entries; extract validation module
Root causes of the 3 failing tests:

1. `skips unselected media entries` and `full-site round-trip` (import-export-
   roundtrip.test.ts): both fixtures used fake ASCII bytes as PNG content
   (`strToU8('keep')`, `Buffer.from('fake-png-bytes')`). The new
   `validateAndSanitizeMediaBytesForImport` gate introduced on this branch
   calls `detectAcceptedMime()` on every staged entry before writing to disk;
   plain ASCII returns null → ArchiveMediaValidationError → 400.
   Fix: update both fixtures to use real PNG magic bytes (8-byte signature ±
   padding to keep `sizeBytes` consistent with DB records).

2. `Module size budgets`: `importArchive.ts` grew to 771 lines (ceiling 700)
   after the new validation class and function were added inline.
   Fix: extract `ArchiveMediaValidationError` and
   `validateAndSanitizeMediaBytesForImport` into the focused new module
   `server/handlers/cms/importArchiveMediaValidation.ts` (97 lines);
   `importArchive.ts` drops to 688 lines. Add the new helper to the
   cms-handlers-capability-gated ALLOWLIST (it's not a route handler,
   same pattern as mediaUpload.ts / svgSanitize.ts).
