+++ src/__tests__/architecture/cms-handlers-capability-gated.test.ts
@@ -58,6 +58,10 @@ const ALLOWLIST: ReadonlyMap<string, string> = new Map([
   // handler (`/me/avatar`, `/media`).
   ['mediaUpload.ts', 'Multipart parse helper called by gated parent handlers.'],
   ['svgSanitize.ts', 'Pure SVG sanitiser called by mediaUpload (itself gated parents); no handlers.'],
+  // Byte-level MIME validation + SVG sanitisation for archive imports. Called
+  // by importArchive.ts which gates the route via requireCapability; this file
+  // contains no request handler itself.
+  ['importArchiveMediaValidation.ts', 'Validation/sanitisation helper called by importArchive.ts (which gates via requireCapability); no handlers.'],
   ['mediaUploadDispatch.ts', 'Storage adapter dispatch called by gated parent handlers.'],
   ['mediaUploadExecutor.ts', 'Filesystem write helper called by gated parent handlers.'],
   ['mediaVariants.ts', 'Variant generation helper called by gated parent handlers.'],
